Explicit check for got_int in vgetorpeek() while loop
Before it used the check inside the for(;;) loop which I think is less
clear.
Also add explicit setting of `exiting` to `true` or `false` in each
branch of the main `if()` statement for the while loop.

I don't believe this would change the behaviour in any way, as there
isn't any way that the initial for loop can be broken out of and not
gotten back into. But it's nice to be explicit.