vgetorpeek swap `break` for `continue`
This is almost useless, it does nothing different, but it does
limit the things someone has to think about when following the code.

It was also pointless before the changes I've made recently, there were
two places where 'c' was set to -1 and we used `break` where we didn't
need to set 'c' to -1, and could have simply used `continue` instead.