Remove superfluous else branch
This branch just set keylen to mp_match_len if mp != NULL.
The previous function called was find_typed_map(), and it's this that
determines the value of mp here.
If this function doesn't return NULL for mp, it is because a map was
found.
In this case keylen was set to mp_match_len in find_typed_map(), and
since mp is not NULL, then check_togglepaste() won't have changed
keylen.