get_key_from_user() was using *cp for two things
Separate the uses into measuring the number of bytes read by inchar()
and returning to vgetorpeek() the next character to use.

The only time that it matters what value vgetorpeek() ends up with for
'c' is when it's breaking out of the for loop. This is only when
get_key_from_user() is returning 1.