Slight tidy-up of an if/else clause
Initially it was in the middle of things, so it needed to specify
continue/break/fall-through etc, but now it's at the end of a function,
and given the use of the function, we can remove most flow-control from
it.
This leaves the if/else clause only updating things as required.