:lnoremap is equivalent to :lmap
As a mental model of things this makes a lot more sense -- if it's a
translation between your keyboard and vim proper, then vim shouldn't be
able to tell the difference between a mapped version and an unmapped
version.
Hence it shouldn't be able to refrain from mapping the resultant keys.